requirements.txt
# SPDX-FileCopyrightText: 2017-2022 Contributors to the OpenSTEF project <openstef@lfenergy.org>
#
# SPDX-License-Identifier: MPL-2.0

ciso8601~=2.3.0 # To speed up influxDB queries
geopy~=2.3.0
openstef~=3.4.87
mlflow-skinny>=3.5.1,<4.0.0
pydantic-settings>=2.1.0,<3.0.0
influxdb-client~=1.36.1
mysql-connector-python>=9.1.0
psycopg2-binary~=2.9.6
PyMySQL>=1.0.2,<2.0.0
PyYAML>=6.0,<7.0.0
requests>=2.28.1,<3.0.0
SQLAlchemy
